dhcpcd: Send unique IAID
The current logic uses the kernel iface index when the name is longer
than 4 octets.  This is not stable -- the iface index changes under
various scenarios, such as toggling airplane mode.

Instead, use the lower 4 octets of the iface hwaddr.  This is the same
logic used in the current dhcpcd code and in the ISC DHCP client.

Note the upstream fix is here:

  http://roy.marples.name/projects/dhcpcd/ci/31ea92e995477728?sbs=1

This is several revisions beyond our current code and it is rumored
that the next major AOSP release will switch to a different DHCP
client, so we won't bother trying to sync up with upstream.

Change-Id: I0663b9bcf3803978adcef6d936ae49e128c7b1f2